credentials/requirements.txt
PyYAML==6.0
requests==2.31.0
python-dotenv==1.0.0
google-cloud-secret-manager==2.16.1
PyGithub==1.76
